+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # 2.x branch
 ## 2.9 branch
 ### 2.9.2
-* fixed the `findActive()` method for `VideosCategoriesTable`.
+* fixed `VideosController::indexByDate()` method;
+* fixed the `findActive()` method for `VideosCategoriesTable`;
+* added tests for `VideosController` and `VideosCategoriesController` classes.
 
 ### 2.9.1
 * fixed the `Categories` alias for `Videos` table;